third_party/WebKit/LayoutTests/accessibility/textarea-line-for-index-expected.txt

This tests that the line number of a character offset is computed correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 0
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 1
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(index) is 2
PASS axArea1.lineForIndex(area1.textLength) is 2
PASS successfullyParsed is true

TEST COMPLETE
